telah selesai belajar php

+++ database/hapus-data.php
@@ -1,17 +1,19 @@
 <?php
-if(isset($_GET['hapus'])) {
+if(!empty($_GET['hapus'])) {
     $hapus = $_GET['hapus'];
-include 'koneksi.php';
+    include 'koneksi.php';
 }
 
-if(!empty($hapus)&&$hapus !="") {
-    $delete = "DELETE FROM roles WHERE id='$hapus'";
-    $sql = mysqli_query($koneksi, $delete);
-    if ($sql){
-        echo header("location: table_rolls.php");
-    } else {
-        echo 'data gagal dihapus' . $hapus;
-        echo mysqli_errno($koneksi);
+$query = mysqli_query($koneksi, "SELECT * FROM roles WHERE id='$hapus'");
+$hasil = mysqli_fetch_array($query);
+if($hasil) {
+    if(!empty($hapus)&&$hapus !="") {
+        $delete = "DELETE FROM roles WHERE id='$hapus'";
+        $sql = mysqli_query($koneksi,$delete);
+        if ($sql){
+            echo "DATA BERHASIL DIHAPUS";
+        }
     }
+}else{
+    echo 'DATA GAGAL';
 }
-?>
